+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Fixes a styling issue on the range `DatePicker` for the month and year types
   where the in-range selected date values did not have the correct DS colors.
+- Fixes the responsive styles related to the image in the `FeaturedContent` component.
 
 ## 2.1.5 (February 5, 2024)